Remove lazy static from domain participant factory
