+++ dds/src/implementation/utils/actor.rs
@@ -271,6 +271,49 @@ impl<A> Actor<A> {
                 "Receiver is guaranteed to exist while actor object is alive. Sending must succeed",
             );
     }
+
+    pub fn send_mail_and_await_reply_blocking<M>(&self, mail: M) -> M::Result
+    where
+        A: MailHandler<M> + Send,
+        M: Mail + Send + 'static,
+        M::Result: Send,
+    {
+        let (response_sender, mut response_receiver) = tokio::sync::oneshot::channel();
+
+        let mut send_result = self
+            .sender
+            .try_send(Box::new(ReplyMail::new(mail, response_sender)));
+        // Try sending the mail until it succeeds. This is done instead of calling a tokio::task::block_in_place because this solution
+        // would be only valid when the runtime is multithreaded. For single threaded runtimes this would still cause a panic.
+        while let Err(receive_error) = send_result {
+            match receive_error {
+                tokio::sync::mpsc::error::TrySendError::Full(mail) => {
+                    send_result = self.sender.try_send(mail);
+                }
+
+                tokio::sync::mpsc::error::TrySendError::Closed(_) => {
+                    panic!("With actor object it should always be possible to send mail")
+                }
+            }
+        }
+
+        // Receive on a try_recv() loop checking for error instead of a call to recv() to avoid blocking the thread. This would not cause
+        // a Tokio runtime panic since it is using an std channel but it could cause a single-threaded runtime to hang and further tasks not
+        // being executed.
+        let mut receive_result = response_receiver.try_recv();
+        while let Err(receive_error) = receive_result {
+            match receive_error {
+                tokio::sync::oneshot::error::TryRecvError::Empty => {
+                    receive_result = response_receiver.try_recv();
+                }
+
+                tokio::sync::oneshot::error::TryRecvError::Closed => {
+                    panic!("With actor object there should always be a reply")
+                }
+            }
+        }
+        receive_result.expect("Receive result should be Ok")
+    }
 }
 
 impl<A> Drop for Actor<A> {
